Split up mouse.rs into smaller chunks
