+++ crates/bevy_input/src/mouse/button.rs
@@ -1,0 +1,9 @@
+/// A button on a mouse device
+#[derive(Debug, Hash, PartialEq, Eq, Clone, Copy)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub enum MouseButton {
+    Left,
+    Right,
+    Middle,
+    Other(u16),
+}
