+++ crates/bevy_input/src/mouse/mod.rs
@@ -1,0 +1,6 @@
+pub mod button;
+pub mod event;
+pub mod system;
+pub mod unit;
+
+pub use crate::mouse::{button::*, event::*, system::*, unit::*};
