+++ crates/bevy_input/src/mouse/unit.rs
@@ -1,0 +1,6 @@
+/// Unit of scroll
+#[derive(Debug, Clone, Copy)]
+pub enum MouseScrollUnit {
+    Line,
+    Pixel,
+}
